Add extra dimension to HDF5 dataset when in Dual Energy
Currently, when saving images in any image mode using `NDPluginHDF5`
the resulting dataset will have shape: `(n_img, height, width)`. When
using Dual Energy, this results in having the images for counter low and
high alternating in the same dimension.

To make the dataset easier to navigate, this commit adds new records to
the IOC that compute and update the extra (virtual) dimension when in
Dual Energy mode. This way we get the same dataset shape as when using
the `pimega-backend` saver.