Feat: Add NumCameraCapture PV
Previously, it wasn’t possible to configure different exposure
counts on the IOC for the detector and the backend. As a result,
whenever an external pulse was received, a full image acquisition
was always performed.

The expected behavior is that external pulses should trigger only
the detector’s NumExposures, while the backend should be allowed
to use a different expected count. This mirrors the existing
separation on the console between camera_numexp and det_numexp,
for which the IOC had no equivalent for camera_numexp.

This change introduces the new NumCameraCapture PV to provide
this separation.